New registration form for Mentors (#833)
* feat: create new design for mentor registry page

* feat: fix bugs related to the new design

* feat: refactoring registry pages
