+++ client/src/modules/Registry/components/index.ts
@@ -1,0 +1,3 @@
+export { GeneralMentor } from './GeneralMentor';
+export { Mentorship } from './Mentorship';
+export { Info } from './Info';
