+++ client/src/modules/Registry/constants/index.ts
@@ -1,0 +1,20 @@
+export const TECHNOLOGIES = [
+  { id: 'nodejs', displayName: 'Node.js' },
+  { id: 'react', displayName: 'React' },
+  { id: 'angular', displayName: 'Angular' },
+  { id: 'ios', displayName: 'IOS' },
+  { id: 'android', displayName: 'Android' },
+];
+
+export const DEFAULT_COLUMN_SIZES = { xs: 24, sm: 24, md: 24, lg: 24, xl: 24 };
+export const DEFAULT_DOUBLE_COLUMN_SIZES = { xs: 24, sm: 24, md: 12, lg: 12, xl: 12 };
+export const DEFAULT_ROW_GUTTER = 24;
+
+export const TEXT_LOCATION_STUDENT_TOOLTIP =
+  'We need your location for understanding the audience and for mentor distribution.';
+export const TEXT_LOCATION_MENTOR_TOOLTIP =
+  'We need your location for understanding audience and use it for students distribution';
+export const TEXT_EMAIL_TOOLTIP = 'No spam e-mails. Only for course purposes.';
+export const TEXT_EPAM_EMAIL_TOOLTIP =
+  'If you are EPAM employee, please specify your email to avoid some manual processes later';
+export const RSSCHOOL_BOT_LINK = 'https://t.me/rsschool_bot?start';
